pyproject.toml
[project]
name = "finance-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.4.0",
    "kiteconnect>=5.0.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "tavily-python>=0.7.10",
    "yfinance>=0.2.65",
]
